+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/)
 ---
 Added
  - discard() and and set_storage() methods for containers
+ - gel_purify instruction to instruction.py and protocol.py
 Changed
 
 Removed